// packages/manager/apps/web-domains/src/domain/enum/domainState.enum.ts
export enum DomainStateEnum {
  DELETED = 'DELETED',
  EXPIRED = 'EXPIRED',
  OK = 'OK',
  PENDING_CREATE = 'PENDING_CREATE',
  PENDING_DELETE = 'PENDING_DELETE',
  PENDING_INTERNAL_TRANSFER = 'PENDING_INTERNAL_TRANSFER',
  PENDING_OUTGOING_TRANSFER = 'PENDING_OUTGOING_TRANSFER',
  RESTORABLE = 'RESTORABLE',
  TO_DELETE = 'TO_DELETE',
}

export enum AdditionalDomainStateEnum {
  DISPUTE = 'DISPUTE',
  TECHNICAL_SUSPENDED = 'TECHNICAL_SUSPENDED',
  OVH_ABUSE = 'OVH_ABUSE',
  DGCCRF_ABUSE = 'DGCCRF_ABUSE',
  REGISTRY_ABUSE = 'REGISTRY_ABUSE',
  FORCED_DELETION = 'FORCED_DELETION',
  PROCEDURE_IN_PROGRESS = 'PROCEDURE_IN_PROGRESS',
}